scripts/loading_co2_data_sp/requirements.txt
snowflake-snowpark-python
python-dotenv
pandas
boto3